+++ extensions/linear/src/hooks/useProjects.ts
@@ -2,20 +2,22 @@ import { useCachedPromise } from "@raycast/utils";
 import type { PaginationOptions } from "@raycast/utils/dist/types";
 
 import { getProjects, type ProjectResult } from "../api/getProjects";
+import { useWorkspaces } from "../components/WorkspaceContext";
 
 export default function useProjects(
   teamId?: string,
   config?: { execute?: boolean; searchText?: string; pageSize?: number },
 ) {
+  const { workspaceKey } = useWorkspaces();
   const { data, error, isLoading, mutate, pagination } = useCachedPromise(
-    (teamId?: string, searchText?: string) => (pagination: PaginationOptions<ProjectResult[]>) =>
+    (key: string, teamId?: string, searchText?: string) => (pagination: PaginationOptions<ProjectResult[]>) =>
       getProjects({
         teamId,
         searchText,
         after: pagination.cursor,
         first: config?.pageSize,
       }),
-    [teamId, config?.searchText],
+    [workspaceKey, teamId, config?.searchText],
     {
       execute: config?.execute !== false,
       keepPreviousData: true,